+++ cmd/psweb/config/common.go
@@ -57,7 +57,7 @@ func Load(dataDir string) {
 	// load defaults first
 	Config.AllowSwapRequests = true
 	Config.ColorScheme = "dark" // dark or light
-	Config.MaxHistory = 10
+	Config.MaxHistory = 20
 	Config.ElementsPass = ""
 	Config.BitcoinSwaps = true
 	Config.LocalMempool = ""
@@ -104,7 +104,7 @@ func Load(dataDir string) {
 		if err != nil {
 			log.Println("Error creating config file.", err)
 		} else {
-			log.Println("Config file created using defaults.")
+			log.Println("Config file created in", Config.DataDir)
 		}
 		return
 	}