+++ cmd/psweb/ln/cln.go
@@ -343,7 +343,60 @@ func CanRBF() bool {
 	return true
 }
 
-// fetch routing statistics for a channel from a given timestamp
+type ListForwardsRequest struct {
+	Status string `json:"status"`
+	Index  string `json:"index"`
+	Start  uint64 `json:"start"`
+}
+
+func (r *ListForwardsRequest) Name() string {
+	return "listforwards"
+}
+
+type Forwarding struct {
+	CreatedIndex uint64  `json:"created_index"`
+	InChannel    string  `json:"in_channel"`
+	OutChannel   string  `json:"out_channel"`
+	OutMsat      uint64  `json:"out_msat"`
+	FeeMsat      uint64  `json:"fee_msat"`
+	ResolvedTime float64 `json:"resolved_time"`
+}
+
+var forwards struct {
+	Forwards []Forwarding `json:"forwards"`
+}
+
+// fetch routing statistics from cln
+func FetchForwardingStats() {
+	// refresh history
+	client, clean, err := GetClient()
+	if err != nil {
+		return
+	}
+	defer clean()
+
+	var newForwards struct {
+		Forwards []Forwarding `json:"forwards"`
+	}
+
+	start := uint64(0)
+	if len(forwards.Forwards) > 0 {
+		// continue from the last index + 1
+		start = forwards.Forwards[len(forwards.Forwards)-1].CreatedIndex + 1
+	}
+
+	// get incremental history
+	client.Request(&ListForwardsRequest{
+		Status: "settled",
+		Index:  "created",
+		Start:  start,
+	}, &newForwards)
+
+	// append to all history
+	forwards.Forwards = append(forwards.Forwards, newForwards.Forwards...)
+}
+
+// get routing statistics for a channel
 func GetForwardingStats(lndChannelId uint64) *ForwardingStats {
 	var (
 		result          ForwardingStats
@@ -361,24 +414,6 @@ func GetForwardingStats(lndChannelId uint64) *ForwardingStats {
 		assistedMsat6m  uint64
 	)
 
-	// refresh history
-	client, clean, err := GetClient()
-	if err != nil {
-		log.Println("GetClient:", err)
-		return &result
-	}
-	defer clean()
-
-	var forwards struct {
-		Forwards []glightning.Forwarding `json:"forwards"`
-	}
-
-	err = client.Request(&glightning.ListForwardsRequest{}, &forwards)
-	if err != nil {
-		log.Println(err)
-		return &result
-	}
-
 	// historic timestamps in sec
 	now := time.Now()
 	timestamp7d := float64(now.AddDate(0, 0, -7).Unix())
@@ -388,33 +423,31 @@ func GetForwardingStats(lndChannelId uint64) *ForwardingStats {
 	channelId := ConvertLndToClnChannelId(lndChannelId)
 
 	for _, e := range forwards.Forwards {
-		if e.Status == "settled" {
-			if e.OutChannel == channelId {
-				if e.ReceivedTime > timestamp6m {
-					amountOut6m += e.MilliSatoshiOut
-					feeMsat6m += e.Fee
-					if e.ReceivedTime > timestamp30d {
-						amountOut30d += e.MilliSatoshiOut
-						feeMsat30d += e.Fee
-						if e.ReceivedTime > timestamp7d {
-							amountOut7d += e.MilliSatoshiOut
-							feeMsat7d += e.Fee
-							log.Println(e)
-						}
+		if e.OutChannel == channelId {
+			if e.ResolvedTime > timestamp6m {
+				amountOut6m += e.OutMsat
+				feeMsat6m += e.FeeMsat
+				if e.ResolvedTime > timestamp30d {
+					amountOut30d += e.OutMsat
+					feeMsat30d += e.FeeMsat
+					if e.ResolvedTime > timestamp7d {
+						amountOut7d += e.OutMsat
+						feeMsat7d += e.FeeMsat
+						log.Println(e)
 					}
 				}
 			}
-			if e.InChannel == channelId {
-				if e.ReceivedTime > timestamp6m {
-					amountIn6m += e.MilliSatoshiOut
-					assistedMsat6m += e.Fee
-					if e.ReceivedTime > timestamp30d {
-						amountIn30d += e.MilliSatoshiOut
-						assistedMsat30d += e.Fee
-						if e.ReceivedTime > timestamp7d {
-							amountIn7d += e.MilliSatoshiOut
-							assistedMsat7d += e.Fee
-						}
+		}
+		if e.InChannel == channelId {
+			if e.ResolvedTime > timestamp6m {
+				amountIn6m += e.OutMsat
+				assistedMsat6m += e.FeeMsat
+				if e.ResolvedTime > timestamp30d {
+					amountIn30d += e.OutMsat
+					assistedMsat30d += e.FeeMsat
+					if e.ResolvedTime > timestamp7d {
+						amountIn7d += e.OutMsat
+						assistedMsat7d += e.FeeMsat
 					}
 				}
 			}
@@ -428,12 +461,71 @@ func GetForwardingStats(lndChannelId uint64) *ForwardingStats {
 	result.AmountIn30d = amountIn30d / 1000
 	result.AmountIn6m = amountIn6m / 1000
 
-	result.FeeSat7d = feeMsat7d
-	result.AssistedFeeSat7d = assistedMsat7d
-	result.FeeSat30d = feeMsat30d
-	result.AssistedFeeSat30d = assistedMsat30d
-	result.FeeSat6m = feeMsat6m
-	result.AssistedFeeSat6m = assistedMsat6m
+	result.FeeSat7d = feeMsat7d / 1000
+	result.AssistedFeeSat7d = assistedMsat7d / 1000
+	result.FeeSat30d = feeMsat30d / 1000
+	result.AssistedFeeSat30d = assistedMsat30d / 1000
+	result.FeeSat6m = feeMsat6m / 1000
+	result.AssistedFeeSat6m = assistedMsat6m / 1000
 
 	return &result
+}
+
+// net balance change for a channel
+func GetNetFlow(lndChannelId uint64, timeStamp uint64) int64 {
+
+	netFlow := int64(0)
+	channelId := ConvertLndToClnChannelId(lndChannelId)
+	timeStampF := float64(timeStamp)
+
+	for _, e := range forwards.Forwards {
+		if e.InChannel == channelId {
+			if e.ResolvedTime > timeStampF {
+				netFlow -= int64(e.OutMsat)
+			}
+		}
+		if e.OutChannel == channelId {
+			if e.ResolvedTime > timeStampF {
+				netFlow += int64(e.OutMsat)
+			}
+		}
+	}
+	return netFlow / 1000
+}
+
+type ListPeerChannelsRequest struct {
+	PeerId string `json:"id,omitempty"`
+}
+
+func (r ListPeerChannelsRequest) Name() string {
+	return "listpeerchannels"
+}
+
+// get fees on the channel
+func GetChannelInfo(client *glightning.Lightning, lndChannelId uint64, nodeId string) *ChanneInfo {
+	info := new(ChanneInfo)
+	channelId := ConvertLndToClnChannelId(lndChannelId)
+
+	var response map[string]interface{}
+
+	err := client.Request(&ListPeerChannelsRequest{
+		PeerId: nodeId,
+	}, &response)
+	if err != nil {
+		log.Println(err)
+		return info
+	}
+
+	// Iterate over channels to find ours
+	channels := response["channels"].([]interface{})
+	for _, channel := range channels {
+		channelMap := channel.(map[string]interface{})
+		if channelMap["short_channel_id"].(string) == channelId {
+			info.FeeBase = uint64(channelMap["fee_base_msat"].(float64))
+			info.FeeRate = uint64(channelMap["fee_proportional_millionths"].(float64))
+			break
+		}
+	}
+
+	return info
 }