+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Versions
 
+## 1.3.4
+
+- Display channel flows since their last swaps
+- Allow filtering swaps history 
+- Speed up loading of pages
+- Display current fee rates for channels
+- Add help tooltips
+- CLN: implement incremental forwards history polling
+- LND 0.18+: exact fee when sending change-less peg-in tx (there was a bug in LND below 0.18)
+
 ## 1.3.3
 
 - Add channel routing statistics on the peer screen